requirements.txt
# Core Dependencies
langchain
langgraph
langchain-core
google-generativeai==0.3.2

# Notion Integration
notion-client==2.2.1

# Vector Database & Embeddings
chromadb==1.0.9

# Database
supabase==2.17.0

# Utilities
python-dotenv>=1.0.1
pydantic>=1.9
pydantic-settings>=2.0.0
python-dateutil>=2.9

# Deployment
google-cloud-scheduler==2.13.0
google-cloud-run==0.10.9 

langsmith>=0.1.8
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
APScheduler>=3.10.4
pytest>=8.1.1

markupsafe>=2.0,<2.1.4

# For vector visualization
scikit-learn
matplotlib
numpy
